refactor: move to `/smu`
